requirements.txt
Flask==2.3.2
requests==2.31.0
SQLAlchemy==2.0.15
pymysql==1.1.0
